+++ src/PhpSpreadsheet/ReferenceHelper.php
@@ -367,7 +367,6 @@ class ReferenceHelper
         Worksheet $worksheet
     ): void {
         $remove = ($numberOfColumns < 0 || $numberOfRows < 0);
-        $allCoordinates = $worksheet->getCoordinates();
 
         if (
             $this->cellReferenceHelper === null ||
@@ -394,12 +393,13 @@ class ReferenceHelper
         }
 
         // Find missing coordinates. This is important when inserting column before the last column
+        $cellCollection = $worksheet->getCellCollection();
         $missingCoordinates = array_filter(
             array_map(function ($row) use ($highestColumn) {
                 return $highestColumn . $row;
             }, range(1, $highestRow)),
-            function ($coordinate) use ($allCoordinates) {
-                return !in_array($coordinate, $allCoordinates);
+            function ($coordinate) use ($cellCollection) {
+                return $cellCollection->has($coordinate) === false;
             }
         );
 
@@ -408,16 +408,15 @@ class ReferenceHelper
             foreach ($missingCoordinates as $coordinate) {
                 $worksheet->createNewCell($coordinate);
             }
-
-            // Refresh all coordinates
-            $allCoordinates = $worksheet->getCoordinates();
         }
 
-        // Loop through cells, bottom-up, and change cell coordinate
+        $allCoordinates = $worksheet->getCoordinates();
         if ($remove) {
             // It's faster to reverse and pop than to use unshift, especially with large cell collections
             $allCoordinates = array_reverse($allCoordinates);
         }
+
+        // Loop through cells, bottom-up, and change cell coordinate
         while ($coordinate = array_pop($allCoordinates)) {
             $cell = $worksheet->getCell($coordinate);
             $cellIndex = Coordinate::columnIndexFromString($cell->getColumn());
@@ -924,31 +923,40 @@ class ReferenceHelper
 
     private function clearColumnStrips(int $highestRow, int $beforeColumn, int $numberOfColumns, Worksheet $worksheet): void
     {
-        for ($i = 1; $i <= $highestRow - 1; ++$i) {
-            for ($j = $beforeColumn - 1 + $numberOfColumns; $j <= $beforeColumn - 2; ++$j) {
-                $coordinate = Coordinate::stringFromColumnIndex($j + 1) . $i;
-                $worksheet->removeConditionalStyles($coordinate);
-                if ($worksheet->cellExists($coordinate)) {
-                    $worksheet->getCell($coordinate)->setValueExplicit('', DataType::TYPE_NULL);
-                    $worksheet->getCell($coordinate)->setXfIndex(0);
-                }
+        $startColumnId = Coordinate::stringFromColumnIndex($beforeColumn + $numberOfColumns);
+        $endColumnId = Coordinate::stringFromColumnIndex($beforeColumn);
+
+        for ($row = 1; $row <= $highestRow - 1; ++$row) {
+            for ($column = $startColumnId; $column !== $endColumnId; ++$column) {
+                $coordinate = $column . $row;
+                $this->clearStripCell($worksheet, $coordinate);
             }
         }
     }
 
     private function clearRowStrips(string $highestColumn, int $beforeColumn, int $beforeRow, int $numberOfRows, Worksheet $worksheet): void
     {
-        $lastColumnIndex = Coordinate::columnIndexFromString($highestColumn) - 1;
+        $startColumnId = Coordinate::stringFromColumnIndex($beforeColumn);
+        ++$highestColumn;
 
-        for ($i = $beforeColumn - 1; $i <= $lastColumnIndex; ++$i) {
-            for ($j = $beforeRow + $numberOfRows; $j <= $beforeRow - 1; ++$j) {
-                $coordinate = Coordinate::stringFromColumnIndex($i + 1) . $j;
-                $worksheet->removeConditionalStyles($coordinate);
-                if ($worksheet->cellExists($coordinate)) {
-                    $worksheet->getCell($coordinate)->setValueExplicit('', DataType::TYPE_NULL);
-                    $worksheet->getCell($coordinate)->setXfIndex(0);
-                }
+        for ($column = $startColumnId; $column !== $highestColumn; ++$column) {
+            for ($row = $beforeRow + $numberOfRows; $row <= $beforeRow - 1; ++$row) {
+                $coordinate = $column . $row;
+                $this->clearStripCell($worksheet, $coordinate);
             }
+        }
+    }
+
+    private function clearStripCell(Worksheet $worksheet, string $coordinate): void
+    {
+        $worksheet->removeConditionalStyles($coordinate);
+        $worksheet->setHyperlink($coordinate);
+        $worksheet->setDataValidation($coordinate);
+        $worksheet->removeComment($coordinate);
+
+        if ($worksheet->cellExists($coordinate)) {
+            $worksheet->getCell($coordinate)->setValueExplicit(null, DataType::TYPE_NULL);
+            $worksheet->getCell($coordinate)->setXfIndex(0);
         }
     }